+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3953,7 +3953,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M0FG0SWD0PPRSXBRPTDN5ZQ4"),
+  id: z.string().default("01M0FN88YTGWAWP1BXXGP5GWB6"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>