+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-20T11:48:20.273Z",
+    "2026-08-20T13:19:48.930Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {