+++ packages/orq-rc/RELEASES.md
@@ -11769,3 +11769,13 @@ Based on:
 - [typescript v4.14.0-rc.40] packages/orq-rc
 ### Releases
 - [NPM v4.14.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.14.0-rc.40 - packages/orq-rc
+
+## 2026-08-20 13:23:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.795.0 (2.931.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.14.0-rc.41] packages/orq-rc
+### Releases
+- [NPM v4.14.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/4.14.0-rc.41 - packages/orq-rc